mocking celery calls.
I've tried, but celery doesn't want to be integration-tested.
